+++ fwoa-tools/package.json
@@ -19,19 +19,20 @@
     "build": "heft build --clean && rushx pkg-json-lint",
     "build-test": "heft test --clean && rushx pkg-json-lint",
     "clean": "rm -rf build/* node_modules/* dist/* .nyc_output/* lib/*",
+    "int-test": "jest --testPathPattern=src/integration-tests/*",
     "lint-fix": "eslint . --fix",
     "local": "node .",
     "pkg-json-lint": "npmPkgJsonLint -c ../.npmpackagejsonlintrc.json .",
     "prettier": "prettier --config '../.prettierrc.js' 'src/**/*.ts' --write",
     "sort-package-json": "sort-package-json package.json",
     "standard-version": "standard-version --skip.tag=true",
-    "test": "heft test --clean --no-build --test-path-pattern src/*.unit.test.ts && rushx pkg-json-lint",
+    "test": "heft test --clean --no-build && rushx pkg-json-lint",
     "watch": "tsc -w"
   },
   "dependencies": {
     "axios": "0.25.0",
-    "object-hash": "^3.0.0",
-    "lodash": "^4.17.21"
+    "lodash": "^4.17.21",
+    "object-hash": "^3.0.0"
   },
   "devDependencies": {
     "@aws/eslint-config-fwoa-eslint-custom": "workspace:*",
