+++ fwoa-tools/src/integration-tests/migrationUtils.test.ts
@@ -2,7 +2,7 @@
  *  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *  SPDX-License-Identifier: Apache-2.0
  */
-import { getFhirClient, getFhirClientSMART } from './migrationUtils';
+import { getFhirClient, getFhirClientSMART } from '../migrationUtils';
 
 describe('MigrationUtils', () => {
   describe('getFhirClient', () => {
